Split pname_to_uid into its own module
Not available on Windows. Include it conditionally to prevent build
errors on Windows.
