+++ gssapi/raw/ext_pname_to_uid.pyx
@@ -1,0 +1,51 @@
+GSSAPI="BASE"  # This ensures that a full module is generated by Cython
+
+from posix.types cimport uid_t
+
+from gssapi.raw.cython_types cimport *
+from gssapi.raw.names cimport Name
+from gssapi.raw.oids cimport OID
+
+from gssapi.raw.misc import GSSError
+
+cdef extern from "python_gssapi_ext.h":
+    OM_uint32 gss_pname_to_uid(OM_uint32 *minor,
+                               const gss_name_t name,
+                               const gss_OID mech_type,
+                               uid_t *uid_out) nogil
+
+
+def pname_to_uid(Name name not None, OID mech=None):
+    """Get the local UID for a GSSAPI name.
+
+    This method determines the local UID associated with a GSSAPI
+    name, optionally for a given mechanism.
+
+    Note:
+        This function is not available on Windows.
+
+    Args:
+        name (Name): the GSSAPI name to map to a local UID
+        mech (~gssapi.OID): the mechanism to use for the mapping
+            (or None for the default)
+
+    Returns:
+        int: the local UID
+
+    Raises:
+        ~gssapi.exceptions.GSSError
+    """
+    cdef gss_OID m = GSS_C_NO_OID
+    if mech is not None:
+        m = &mech.raw_oid
+
+    cdef uid_t uid_out
+    cdef OM_uint32 maj_stat, min_stat
+
+    with nogil:
+        maj_stat = gss_pname_to_uid(&min_stat, name.raw_name, m, &uid_out)
+
+    if maj_stat == GSS_S_COMPLETE:
+        return uid_out
+    else:
+        raise GSSError(maj_stat, min_stat)
